Minimal Kernel
In this section, two file, kernel.ld and kernel.c, are created.
kernel.ld is a linker script which is to specify the memory
layout of executables file. Based on the layout, the linker
assigns memory addresses to functions and variables. kernel.c
is a minimal kernel which used to boot and initialize the bss
and stack pointer. kernel.c uses the variable declared in the
kernel.ld by using extern. Also, run.sh is modified to build
and compile the source code.